Give the cutoff its own type, which is what its eight dependencies were saying
The decorator had grown to eight constructor parameters, which the
static analysis flagged and which was the honest reading of what it had
become: two questions answered from two stores. What is recorded about
this one token, and what is recorded about the principal behind it. The
second needs the issuer, the client store, the pairwise converter and a
clock tolerance; the first needs none of them.

RevocationCutoffChecker owns the second question, and the decorator asks
it. Nothing about the decision changed - the same tests drive the same
inputs to the same outcomes - but each half can now be read without the
other, and the cutoff can be exercised without a token registry and an
inner validator standing in front of every case.

The unit tests build the real checker over the same mocked stores rather
than a stub of it. A stub told to return the right answer would measure
nothing about the decision these tests exist for.
